Move some feature variations logic to fontir
We want to be able to call the 'overlay_feature_variations' method from
both fontbe and glyphs2fontir.
