+++ fontir/src/lib.rs
@@ -1,6 +1,7 @@
 //! Intermediate Representation (IR) types for font compilation
 
 pub mod error;
+pub mod feature_variations;
 pub mod glyph;
 pub mod ir;
 pub mod orchestration;
